Added common folder to src. Moved Header, Footer, and Button to common
