+++ frontend/src/components/common/Footer/Footer.tsx
