+++ frontend/src/components/common/Footer/footer.styles.ts
