+++ frontend/src/components/common/Header/header.styles.ts
